refactor!: docxstamper classes into officestamper.core package
In this commit, several org.wickedsource.docxstamper.el classes were moved into the pro.verron.officestamper.core package. These include StandardMethodResolver, StandardMethodExecutor, and ExpressionResolver. This restructuring also involved updating import statements in PlaceholderReplacer, CommentProcessorRegistry, and DocxStamper classes. The org.wickedsource.docxstamper.el package was subsequently removed from module-info.java.
