+++ engine/src/main/java/module-info.java
@@ -54,7 +54,5 @@ module pro.verron.officestamper {
     exports pro.verron.officestamper.experimental to pro.verron.officestamper.test;
 
     // TODO_LATER: remove all the following exports in next version
-    exports org.wickedsource.docxstamper.el;
     exports org.wickedsource.docxstamper.util;
-
 }
